pyproject.toml
[tool.poetry]
name = "devops_tools"
version = "0.1.3"
description = "devops tools mostly in python, with tests examples"
authors = ["Nicolas Heim"]
readme = "README.md"

packages = [
    # { include = "**", from = "./src", to = "devops-tools" },
    # { include = "**", to = "devops-tools" },
    { include = "devops_tools" }
]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.10.3"
python-dotenv = "^1.0.1"
hvac = "^2.3.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
